Further refactoring of integration tests
